requirements.txt
uvicorn
fastapi
jinja2
httpx
python-dotenv
openai
markdown2
python-multipart
pdfkit